Add TS, and rehype to parse md files
