+++ components/layout.tsx
@@ -7,7 +7,13 @@ import Link from 'next/link';
 const name = 'Roman K.';
 export const siteTitle = 'Next.js Sample Website';
 
-export default function Layout({ children, home }) {
+export default function Layout({
+  children,
+  home,
+}: {
+  children: React.ReactNode;
+  home?: boolean;
+}) {
   return (
     <div className={styles.container}>
       <Head>
