--- pages/_app.js
@@ -1,5 +1,0 @@
-import '../styles/global.css';
-
-export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
