--- pages/api/hello.js
@@ -1,3 +1,0 @@
-export default function handler(req, res) {
-  res.status(200).json({ text: 'Hello' });
-}
